requirements.txt
# WeChat UI Automation (微信 UI 自动化)
# 支持所有微信版本 (基于 UI 自动化)
pywinauto>=0.6.8
pyautogui>=0.9.53
pillow>=10.0.0

# OCR (屏幕文字识别)
# 可选：使用 Windows 自带 OCR 或 pytesseract
pytesseract>=0.3.10
screen-ocr>=0.7.0

# LLM 相关 - 使用宽松版本让 pip 自动解决冲突
torch>=2.0.0
transformers>=4.30.0,<4.50.0  # 避免与 regex 冲突
peft>=0.5.0
accelerate>=0.20.0,<1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0,<4.0.0  # 避免与 huggingface_hub 冲突

# 工具库
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.25.0  # API 调用

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # 覆盖率报告

# 代码质量
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
